+++ api/managed_account/managed_account_test.go
@@ -13,6 +13,7 @@ import (
 	"time"
 
 	"github.com/BeyondTrust/go-client-library-passwordsafe/api/authentication"
+	"github.com/BeyondTrust/go-client-library-passwordsafe/api/constants"
 	"github.com/BeyondTrust/go-client-library-passwordsafe/api/entities"
 	"github.com/BeyondTrust/go-client-library-passwordsafe/api/logging"
 	"github.com/BeyondTrust/go-client-library-passwordsafe/api/utils"
@@ -41,7 +42,7 @@ type CreateManagedAccountsResponse struct {
 
 // the recommended version is 3.1. If no version is specified,
 // the default API version 3.0 will be used
-var apiVersion string = "3.1"
+var apiVersion string = constants.ApiVersion31
 
 var authParams *authentication.AuthenticationParametersObj
 var zapLogger *logging.ZapLogger
@@ -60,10 +61,10 @@ func InitializeGlobalConfig() {
 	authParams = &authentication.AuthenticationParametersObj{
 		HTTPClient:                 *httpClientObj,
 		BackoffDefinition:          backoffDefinition,
-		EndpointURL:                "https://fake.api.com:443/BeyondTrust/api/public/v3/",
+		EndpointURL:                constants.FakeApiUrl,
 		APIVersion:                 apiVersion,
-		ClientID:                   "fakeone_a654+9sdf7+8we4f",
-		ClientSecret:               "fakeone_a654+9sdf7+8we4f",
+		ClientID:                   constants.FakeClientId,
+		ClientSecret:               constants.FakeClientSecret,
 		ApiKey:                     "",
 		Logger:                     zapLogger,
 		RetryMaxElapsedTimeSeconds: 300,